osint-website/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.25.0
pyyaml==6.0